Move files in the airline original project to the "airline" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
