+++ projects-parent/originals-parent/airline/src/it/java/edu/pdx/cs410J/airline/Project1IT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.airline;
 
 import edu.pdx.cs410J.InvokeMainTestCase;
 import org.junit.Test;
