+++ projects-parent/originals-parent/airline/src/test/java/edu/pdx/cs410J/airline/FlightTest.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.airline;
 
 import org.junit.Test;
